Feat: vérifications et permissions

+++ src/Service/UserService.php
@@ -2,11 +2,11 @@
 
 namespace App\Services;
 
-use App\Enum\UserRoles;
+
 use App\Repositories\UserRelationsRepository;
 use App\Repositories\CarRepository;
 use App\Models\User;
-use App\Services\BaseService;
+use App\Enum\UserRoles;
 
 
 use InvalidArgumentException;
@@ -25,7 +25,7 @@ class UserService extends BaseService
 
     //----------Action VISITEUR------------
 
-    // Permet à un visiteur de créer un compte
+    // Permet à un visiteur de créer un compte.
     public function createAccount(array $data): User
     {
         // Vérifier que l'email n'est pas déjà utilisé
@@ -52,9 +52,9 @@ class UserService extends BaseService
         $user->setUserZipCode($data['zip_code']);
         $user->setUserUriPicture($data['picture']) ?? null;
         $user->setUserCredits(($data['credits'] ?? 0) + 20);
-        // Le rôle passager est défini par défaut dans l'entité User
+        $user->setUserRoles([UserRoles::PASSAGER]);
 
-        //Enregistrement dans la BD
+        // Enregistrement dans la BD
         $this->userRelationsRepository->insertUser($user);
 
         return $user;
@@ -62,6 +62,7 @@ class UserService extends BaseService
 
     // -----------------Actions PASSAGER ------------------
 
+    // Permet à un utilisateur de devenir CONDUCTEUR.
     public function becomeDriver(array $data, int $passengerId): User
     {
         // Récupération du passager
@@ -93,19 +94,28 @@ class UserService extends BaseService
     // Permet à un utilisateur PASSAGER OU CONDUCTEUR de supprimer son compte.
     public function deleteAccount(int $userId) //: bool
     {
-        // Vérification des permissions
-        $this->ensurePassenger($userId);
-
-        if ($this->ensureAdmin($userId) || $this->ensureEmployee($userId)) {
-            throw new InvalidArgumentException("Un admin ou un employé ne peut pas supprimer son compte.");
-        }
-
         // Récupération de l'utilisateur
         $user = $this->userRelationsRepository->findUserById($userId);
 
         // Vérification de l'existence de l'utilisateur
         if (!$user) {
             throw new InvalidArgumentException("Utilisateur introuvable.");
+        }
+
+        // Vérification que l'utilisateur a supprimer n'a pas le rôle admin ou employé
+        if (
+            $this->roleService->hasRole($userId, 'EMPLOYE') &&
+            $this->roleService->hasRole($userId, 'ADMIN')
+        ) {
+            throw new InvalidArgumentException("Un admin ou un employé ne peut pas supprimer son compte.");
+        }
+
+        // Vérification des permissions pour PASSAGER et CONDUCTEUR
+        if (
+            !$this->roleService->hasRole($userId, 'PASSAGER') &&
+            !$this->roleService->hasRole($userId, 'CONDUCTEUR')
+        ) {
+            throw new InvalidArgumentException("Seuls les passagers et les conducteurs peuvent supprimer son compte.");
         }
 
         // Enregistrement en Bd
@@ -141,34 +151,46 @@ class UserService extends BaseService
 
 
         // Vérifications des champs modifiés et ajouts des nouvelles valeurs
-        if (isset($newData['last_name'])) {
+        if (!empty($newData['last_name'])) {
             $user->setUserLastName($newData['last_name']);
         }
-        if (isset($newData['first_name'])) {
+        if (!empty($newData['first_name'])) {
             $user->setUserFirstName($newData['first_name']);
         }
-        if (isset($newData['email'])) {
+        if (!empty($newData['email'])) {
+
+            $existingUser = $this->userRelationsRepository->findUserByEmail($newData['email']);
+
+            if ($existingUser && $newData['email'] !== $user->getUserEmail()) {
+                throw new InvalidArgumentException("L'email est déjà utilisé.");
+            }
+
             $user->setUserEmail($newData['email']);
         }
-        if (isset($newData['login'])) {
+        if (!empty($newData['login'])) {
+            $existingUser = $this->userRelationsRepository->findUserByLogin($newData['login']);
+
+            if ($existingUser && $newData['login'] !== $user->getUserLogin()) {
+                throw new InvalidArgumentException("Le login est déjà utilisé.");
+            }
             $user->setUserLogin($newData['login']);
         }
-        if (isset($newData['phone'])) {
+        if (!empty($newData['phone'])) {
             $user->setUserPhone($newData['phone']);
         }
-        if (isset($newData['address'])) {
+        if (!empty($newData['address'])) {
             $user->setUserAddress($newData['address']);
         }
-        if (isset($newData['city'])) {
+        if (!empty($newData['city'])) {
             $user->setUserCity($newData['city']);
         }
-        if (isset($newData['zip_code'])) {
+        if (!empty($newData['zip_code'])) {
             $user->setUserZipCode($newData['zip_code']);
         }
-        if (isset($newData['picture'])) {
+        if (!empty($newData['picture'])) {
             $user->setUserUriPicture($newData['picture']);
         }
-        if (isset($newData['licence_no'])) {
+        if (!empty($newData['licence_no']) && $this->roleService->hasRole($userId, 'CONDUCTEUR')) {
             $user->setUserLicenceNo($newData['licence_no']);
         }
 
@@ -198,16 +220,21 @@ class UserService extends BaseService
             throw new InvalidArgumentException("L'utilisateur n'est pas autorisé à modifier son profil.");
         }
 
+        // Vérification du hashage de l'ancien mot de passe
+        if (!password_verify($oldPassword, $user->getUserPassword())) {
+            throw new InvalidArgumentException("L'ancien mot de passe est incorrect.");
+        }
 
         // Vérification du mot de passe
-        if ($newPassword = $oldPassword) {
+        if ($newPassword === $oldPassword) {
             throw new InvalidArgumentException("Le nouveau mot de passe doit être différent de l'ancien.");
         }
 
-        // Récupération de l'utilisateur
-        $user = $this->userRelationsRepository->findUserById($userId);
-
+        // Mise à jour - hashage inclus dans le setter
         $user->setUserPassword($newPassword);
+
+        // Enregistrement dans la BD
+        $this->userRelationsRepository->updateUser($user);
     }
 
     //----------Actions ADMIN ------------
@@ -215,8 +242,6 @@ class UserService extends BaseService
     // Permet à un admin de créer un compte employé.
     public function createEmployeeAccount(array $data, int $adminId) // : ?User
     {
-
-
         // Récupération de l'admin
         $admin = $this->userRelationsRepository->findUserById($adminId);
 
@@ -227,25 +252,37 @@ class UserService extends BaseService
 
         //Verification de la permission
         $this->ensureAdmin($adminId);
+
+
+        // Vérifier que l'email n'est pas déjà utilisé
+        $existingUser = $this->userRelationsRepository->findUserByEmail($data['email']);
+
+        if ($existingUser) {
+            throw new InvalidArgumentException("Cet email est déjà utilisé.");
+        }
+
+
+        //Création de l'objet User vide
+        $user = new User(
+            $data['last_name'],
+            $data['first_name'],
+            $data['email'],
+            $data['password'],
+            false
+        );
+
+        // Définition du rôle EMPLOYE
+        $user->setUserRoles([UserRoles::EMPLOYE]);
+
+        // Enregistrement dans la BD
+        $this->userRelationsRepository->insertUser($user);
+
+        return $user;
     }
 
     // Permet à un admin de supprimer n'importe quel compte.
     public function deleteAccountByAdmin(int $adminId, int $userId): void
     {
-
-
-
-        // Récupération de l'admin
-        $admin = $this->userRelationsRepository->findUserById($adminId);
-
-        // Vérification de l'existence de l'admin
-        if (!$admin) {
-            throw new InvalidArgumentException("Utilisateur introuvable.");
-        }
-
-        //Verification de la permission
-        $this->ensureAdmin($adminId);
-
         // Récupération de l'utilisateur
         $user = $this->userRelationsRepository->findUserById($userId);
 
@@ -254,46 +291,20 @@ class UserService extends BaseService
             throw new InvalidArgumentException("Utilisateur introuvable.");
         }
 
+
+        // Récupération de l'admin
+        $admin = $this->userRelationsRepository->findUserById($adminId);
+
+        // Vérification de l'existence de l'admin
+        if (!$admin) {
+            throw new InvalidArgumentException("Utilisateur introuvable.");
+        }
+
+        //Verification de la permission
+        $this->ensureAdmin($adminId);
+
+
         // Enregistrement en Bd
         $this->userRelationsRepository->deleteUser($userId);
-    }
-
-
-
-
-
-    // --------------------------------------A METTRE DANS REVIEWSERVICE
-    // Permet à un utilisateur PASSAGER de laisser un commentaire à un utilisateur CONDUCTEUR.
-    public function leaveReview(int $passengerId): void
-    {
-
-        // Récupération du passager
-        $passenger = $this->userRelationsRepository->findUserById($passengerId);
-
-        // Vérification de l'existence du passeger
-        if (!$passenger) {
-            throw new InvalidArgumentException("Passager introuvable.");
-        }
-
-        // Vérification de la permission
-        $this->ensurePassenger($passengerId);
-    }
-
-    //----------Actions EMPLOYEE ------------
-
-    //Fonction qui permet à un employé d'approuver un commentaire.
-    public function approveReview(int $employeeId, string $review, int $rate): void
-    {
-
-        // Récupération de l'employé
-        $employee = $this->userRelationsRepository->findUserById($employeeId);
-
-        // Vérification de l'existence de l'employé
-        if (!$employee) {
-            throw new InvalidArgumentException("Employé introuvable.");
-        }
-
-        // Vérification de la permission
-        $this->ensureEmployee($employeeId);
     }
 }
